requirements.txt
asgiref==3.8.1
Django==5.1.4
django-dotenv==1.4.2
django-querycount==0.8.3
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2024.2
